geniectl/pyproject.toml
[project]
name = "geniectl"
version = "0.7.0"
description = "A tool to declaratively generate multimedia assets using Kubernetes-like YAML files."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["ai", "llm", "generative-ai", "kubernetes", "yaml", "media", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click",
    "pyyaml",
    "google-generativeai",
    "requests",
]

[project.scripts]
geniectl = "geniectl.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "build",
    "twine",
]

[tool.pytest.ini_options]
python_files = "*_test.py"
